fix(simplex): close correlated lifecycle review gaps
Integrate and strengthen the four independently reported SimpleX fixes: lossless cancellation recovery, multiplex profile isolation, daemon listener locking, and allowed-group authorization propagation. Split the adapter into bounded plugin-owned modules while preserving its public surface.

Integrated-from: b32aba0ceb15969f6648c4ef0828b8d8a905aa6b
Integrated-from: 32b78ef80d5667467feee72d0a3089aa997a2634
Integrated-from: 3a3ddf5186d34c88870180c94e69f9a14b5f51e3
Integrated-from: 2d3081f9eb97b5f4b6f1cdf1d895e41009798cf1
